+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Logged out users not loading `check-permission`
 ## [0.1.1] - 2021-08-13
 ### Fixed
 - Not loading `checkUserPermission`